+++ node/eslint.config.js
@@ -1,37 +1,24 @@
-import js_plugin from "@eslint/js";
-import ts_plugin from "@typescript-eslint/eslint-plugin";
+// @ts-check
+import js_eslint from "@eslint/js";
+import ts_eslint from 'typescript-eslint';
 
-import ts_parser from "@typescript-eslint/parser";
-
-// REF https://eslint.nodejs.cn/docs/latest/use/configure/configuration-files-new
-export default [
-    {
-        ignores: [
-            "./coverage/**",
-            "./dist/**",
-            "./node_modules/**",
-            "./temp/**",
-        ],
-    },
-    {
-        files: [
-            "**/*.js",
-            "**/*.ts",
-            "**/*.d.ts",
-        ],
-        languageOptions: {
-            parser: ts_parser,
-            parserOptions: {
-                ecmaVersion: "latest",
-                sourceType: "module",
+// REF: https://eslint.nodejs.cn/docs/latest/use/configure/configuration-files-new
+// REF: https://typescript-eslint.io/getting-started/#step-2-configuration
+export default ts_eslint.config(
+    js_eslint.configs.recommended,
+    ...ts_eslint.configs.recommended,
+    ...[
+        {
+            ignores: [
+                "./*.cjs",
+                "./coverage/**",
+                "./dist/**",
+                "./node_modules/**",
+                "./temp/**",
+            ],
+            // REF https://eslint.nodejs.cn/docs/latest/rules/
+            rules: {
             },
         },
-        plugins: {
-            "eslint:recommended": js_plugin,
-            "@typescript-eslint": ts_plugin,
-        },
-        // REF https://eslint.nodejs.cn/docs/latest/rules/
-        rules: {
-        },
-    },
-];
+    ],
+);