+++ node/src/types/manifest/theme.d.ts
@@ -20,17 +20,50 @@
  * Theme resource manifest file definition
  */
 export interface ITheme {
-    author:         string;
-    description?:   Description;
-    displayName?:   Description;
-    funding?:       Funding;
-    keywords?:      string[];
+    /**
+     * The modes supported by the theme
+     */
+    modes?: TMode[];
+    /**
+     * The name of the author
+     */
+    author: string;
+    /**
+     * The description of the resource
+     */
+    description?: ILocalizedText;
+    /**
+     * The display name of the resource
+     */
+    displayName?: ILocalizedText;
+    /**
+     * The funding of the resource
+     */
+    funding?: IFunding;
+    /**
+     * The keywords of the resource used for search
+     */
+    keywords?: string[];
+    /**
+     * The minimum version of SiYuan that the resource is compatible with
+     */
     minAppVersion?: string;
-    modes?:         IThem[];
-    name:           string;
-    readme?:        Description;
-    url:            string;
-    version:        string;
+    /**
+     * The name of the resource
+     */
+    name: string;
+    /**
+     * The readme file name of the resource
+     */
+    readme?: ILocalizedText;
+    /**
+     * The GitHub repository URL of the resource
+     */
+    url: string;
+    /**
+     * The version of the resource
+     */
+    version: string;
     [property: string]: any;
 }
 
@@ -43,7 +76,7 @@ export interface ITheme {
  *
  * The readme file name of the resource
  */
-export interface Description {
+export interface ILocalizedText {
     /**
      * The default text
      */
@@ -66,7 +99,7 @@ export interface Description {
 /**
  * The funding of the resource
  */
-export interface Funding {
+export interface IFunding {
     /**
      * The custom funding URLs
      */
@@ -86,10 +119,8 @@ export interface Funding {
 }
 
 /**
- * The modes supported by the theme
- *
  * The mode supported by the theme
  */
-export type IThem = "dark" | "light";
+export type TMode = "light" | "dark";
 
 //#endregion content